Performance improvement: omit transformation if it's the identity matrix
In my project, which generates thousands of svg in a very short amount
of time, transformDP is a major performance killer. This is a pitty
because I do not use transformation matrices (I precompute the
transformations and generate the shapes with the right coordinates
directly).

This patch greatly helps the performance my code, with no rendering
difference for my svg files. In addition, the svg files are much
smaller too.